Most properties of the type int are converted to long.

+++ RabbitMqHttpApiClient/Models/Common/MessageStatsModel/IConfirm.cs
@@ -5,7 +5,7 @@ namespace RabbitMqHttpApiClient.Models.Common.MessageStatsModel
     public interface IConfirm
     {
         [JsonProperty("confirm")]
-        int Confirm { get; set; }
+        long Confirm { get; set; }
 
         [JsonProperty("confirm_details")]
         RateDetails ConfirmDetails { get; set; }
