+++ RabbitMqHttpApiClient/Models/Common/MessageStatsModel/IReturnUnroutable.cs
@@ -5,7 +5,7 @@ namespace RabbitMqHttpApiClient.Models.Common.MessageStatsModel
     public interface IReturnUnroutable
     {
         [JsonProperty("return_unroutable")]
-        int ReturnUnroutable { get; set; }
+        long ReturnUnroutable { get; set; }
 
         [JsonProperty("return_unroutable_details")]
         RateDetails ReturnUnroutableDetails { get; set; }
