+++ RabbitMqHttpApiClient/Models/QueueModel/BackingQueueStatus.cs
@@ -8,28 +8,28 @@ namespace RabbitMqHttpApiClient.Models.QueueModel
         public string Mode { get; set; }
 
         [JsonProperty("q1")]
-        public int Q1 { get; set; }
+        public long Q1 { get; set; }
 
         [JsonProperty("q2")]
-        public int Q2 { get; set; }
+        public long Q2 { get; set; }
 
         [JsonProperty("delta")]
         public object[] Delta { get; set; }
 
         [JsonProperty("q3")]
-        public int Q3 { get; set; }
+        public long Q3 { get; set; }
 
         [JsonProperty("q4")]
-        public int Q4 { get; set; }
+        public long Q4 { get; set; }
 
         [JsonProperty("len")]
-        public int Len { get; set; }
+        public long Len { get; set; }
 
         [JsonProperty("target_ram_count")]
         public string TargetRamCount { get; set; }
 
         [JsonProperty("next_seq_id")]
-        public int NextSeqId { get; set; }
+        public long NextSeqId { get; set; }
 
         [JsonProperty("avg_ingress_rate")]
         public double AvgIngressRate { get; set; }
